Work on the Kmymoney importer

+++ alere_lib/src/scenarios.rs
@@ -1,3 +1,3 @@
-pub type Scenario = u16;
+pub type Scenario = i32;
 
 pub const NO_SCENARIO: Scenario = 0;
